pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vocal-scriber"
version = "0.1.0"
description = "Push-to-talk voice typing for terminals on macOS and Windows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "sounddevice>=0.4.6",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "requests>=2.28.0",
    "screeninfo>=0.8.1",
]

[project.optional-dependencies]
macos = [
    "faster-whisper>=0.10.0",
    "pyobjc-framework-Cocoa>=10.0",
]
windows = [
    "openai-whisper>=20231117",
    "pyautogui>=0.9.53",
    "pystray>=0.19.5",
    "pillow>=10.0.0",
]
windows-build = [
    "pyinstaller>=6.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
